added A TON of functions.
html code is better organized. still not commented (:
You can now gracefully cancel uploads, while they're being uploaded. Will throw an appropriate error on device as well.
You can see the file that is being uploaded with a spinning circle in the log. No percentage yet but it's a start.
Reversed the log history because that makes more sense.